refactor(): contact us view model
