+++ CleanCodeApp/Modules/Features/Luz/ContactsUs/LuzContactUsFactory.swift
@@ -1,0 +1,9 @@
+import UIKit
+
+enum LuzContactUsFactory {
+    static func make() -> UIViewController {
+        let service = ContactUSService()
+        let viewModel = LuzContactUSViewModel(serivce: service)
+        return LuzContactUsViewController(viewModel: viewModel)
+    }
+}
