+++ CleanCodeApp/Modules/Features/Luz/ResetPassword/LuzResetPasswordViewController.swift
@@ -60,7 +60,7 @@ final class LuzResetPasswordViewController: UIViewController {
     }
 
     @IBAction func helpDidTap(_ sender: Any) {
-        let vc = LuzContactUsViewController()
+        let vc = LuzContactUsFactory.make()
         vc.modalPresentationStyle = .fullScreen
         vc.modalTransitionStyle = .coverVertical
         self.present(vc, animated: true, completion: nil)
